+++ core/components/crosscontextssettings/lexicon/en/setting.inc.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Setting Lexicon Entries for CrossContextsSettings
+ * Setting lexicon entries for CrossContextsSettings
  *
  * @package crosscontextssettings
  * @subpackage lexicon