+++ core/components/crosscontextssettings/processors/mgr/contextssettings/getlist.class.php
@@ -1,35 +1,22 @@
 <?php
 /**
- * Get contexts settings list processor for CrossContextsSettings
+ * Get list contexts settings
  *
  * @package crosscontextssettings
- * @subpackage processor
+ * @subpackage processors
  */
 
-class CrossContextsSettingsSettingsGetListProcessor extends modObjectGetListProcessor
+use TreehillStudio\CrossContextsSettings\Processors\ObjectGetListProcessor;
+
+class CrossContextsSettingsSettingsGetListProcessor extends ObjectGetListProcessor
 {
     public $classKey = 'modContextSetting';
-    public $languageTopics = array('crosscontextssettings:default');
     public $defaultSortField = 'key';
     public $defaultSortDirection = 'ASC';
-    public $objectType = 'crosscontextssettings.settings';
+    public $permission = 'settings';
+    public $objectType = 'setting';
 
-    protected $columns = array();
-    /** @var CrossContextsSettings $crosscontextssettings */
-    protected $crosscontextssettings;
-
-    /**
-     * {@inheritDoc}
-     * @param modX $modx A reference to the modX instance
-     * @param array $properties An array of properties
-     */
-    function __construct(modX &$modx, array $properties = [])
-    {
-        parent::__construct($modx, $properties);
-
-        $corePath = $this->modx->getOption('crosscontextssettings.core_path', null, $this->modx->getOption('core_path') . 'components/crosscontextssettings/');
-        $this->crosscontextssettings = $this->modx->getService('crosscontextssettings', 'CrossContextsSettings', $corePath . 'model/crosscontextssettings/');
-    }
+    protected $columns = [];
 
     /**
      * {@inheritDoc}
@@ -57,65 +44,36 @@ class CrossContextsSettingsSettingsGetListProcessor extends modObjectGetListProc
 
     /**
      * {@inheritDoc}
-     * @return array
-     */
-    public function getData(): array
-    {
-        $data = array();
-        $limit = intval($this->getProperty('limit'));
-        $start = intval($this->getProperty('start'));
-
-        /* query for chunks */
-        $c = $this->modx->newQuery($this->classKey);
-        $c = $this->prepareQueryBeforeCount($c);
-        $data['total'] = $this->modx->getCount($this->classKey, $c);
-        $c = $this->prepareQueryAfterCount($c);
-        $sortClassKey = $this->getSortClassKey();
-        $sortKey = $this->modx->getSelectColumns($sortClassKey, $this->getProperty('sortAlias', $sortClassKey), '', array($this->getProperty('sort')));
-        if (empty($sortKey)) {
-            $sortKey = $this->getProperty('sort');
-        }
-        $c->sortby($sortKey, $this->getProperty('dir'));
-        if ($limit > 0) {
-            $c->limit($limit, $start);
-        }
-
-        $data['results'] = $this->modx->getCollection($this->classKey, $c);
-        return $data;
-    }
-
-    /**
-     * {@inheritDoc}
      * @param xPDOQuery $c
      * @return xPDOQuery
      */
-    public function prepareQueryBeforeCount(xPDOQuery $c): xPDOQuery
+    public function prepareQueryBeforeCount(xPDOQuery $c)
     {
-        $c->select($this->modx->getSelectColumns($this->classKey, $this->classKey, '', array('key')));
-        $c->select('MIN('.$this->classKey.'.context_key) as context_key');
-        $c->where(array(
+        $c->select($this->modx->getSelectColumns($this->classKey, $this->classKey, '', ['key']));
+        $c->select('MIN(' . $this->classKey . '.context_key) as context_key');
+        $c->where([
             $this->classKey . '.context_key:!=' => 'mgr'
-        ));
-        if ($contexts = $this->crosscontextssettings->getOption('contexts')) {
-            $c->where(array($this->classKey . '.context_key:IN' => $contexts));
+        ]);
+        if ($contextsKeys = $this->crosscontextssettings->getOption('contexts_keys')) {
+            $c->where([$this->classKey . '.context_key:IN' => $contextsKeys]);
         }
         $namespace = $this->getProperty('namespace');
         if ($namespace && $namespace != 'core') {
-            $c->where(array(
+            $c->where([
                 $this->classKey . '.namespace:=' => $namespace
-            ));
+            ]);
         }
         $area = $this->getProperty('area');
         if ($area) {
-            $c->where(array(
+            $c->where([
                 $this->classKey . '.area:=' => $area
-            ));
+            ]);
         }
         $key = $this->getProperty('key');
         if ($key) {
-            $c->where(array(
+            $c->where([
                 $this->classKey . '.key:LIKE' => '%' . $key . '%'
-            ));
+            ]);
         }
         $c->groupby($this->classKey . '.key', 'asc');
 
@@ -131,16 +89,16 @@ class CrossContextsSettingsSettingsGetListProcessor extends modObjectGetListProc
     {
         $objectArray = $object->toArray();
 
-        $output = array();
+        $output = [];
         if (!empty($this->columns) && is_array($this->columns)) {
             foreach ($this->columns as $column) {
                 if ($column === 'key' || $column === 'namespace' || $column === 'area') {
                     $output[$column] = $objectArray[$column];
                 } else {
-                    $setting = $this->modx->getObject($this->classKey, array(
+                    $setting = $this->modx->getObject($this->classKey, [
                         'context_key' => $column,
                         'key' => $objectArray['key'],
-                    ));
+                    ]);
                     if (!isset($output['xtype'])) {
                         $output['xtype'] = '';
                     }
@@ -150,10 +108,10 @@ class CrossContextsSettingsSettingsGetListProcessor extends modObjectGetListProc
                     } else {
                         $output[$column] = '';
                         if (empty($output['xtype'])) {
-                            $check = $this->modx->getObject($this->classKey, array(
+                            $check = $this->modx->getObject($this->classKey, [
                                 'key' => $objectArray['key'],
                                 'xtype:!=' => '',
-                            ));
+                            ]);
                             if ($check) {
                                 $output['xtype'] = $check->get('xtype');
                             }