+++ source/js/mgr/crosscontextssettings.js
@@ -1,11 +1,14 @@
 var crosscontextssettings = function (config) {
     config = config || {};
-    Ext.applyIf(config, {});
     crosscontextssettings.superclass.constructor.call(this, config);
-    return this;
 };
 Ext.extend(crosscontextssettings, Ext.Component, {
-    page: {}, window: {}, grid: {}, tree: {}, panel: {}, combo: {}, config: {}
+    initComponent: function () {
+        this.stores = {};
+        this.ajax = new Ext.data.Connection({
+            disableCaching: true,
+        });
+    }, page: {}, window: {}, grid: {}, tree: {}, panel: {}, combo: {}, config: {}, util: {}, form: {}
 });
 Ext.reg('crosscontextssettings', crosscontextssettings);